+++ src/connector/factory.go
@@ -2,10 +2,10 @@ package connector
 
 import (
 	"fmt"
+	"github.com/cranemont/iris/lib/logger"
 
 	"github.com/cranemont/iris/src/connector/rabbitmq"
 	"github.com/cranemont/iris/src/router"
-	"github.com/cranemont/iris/src/service/logger"
 )
 
 type Providers struct {
