+++ src/connector/rabbitmq/connector.go
@@ -3,10 +3,10 @@ package rabbitmq
 import (
 	"context"
 	"fmt"
+	"github.com/cranemont/iris/lib/logger"
 	"time"
 
 	"github.com/cranemont/iris/src/router"
-	"github.com/cranemont/iris/src/service/logger"
 	amqp "github.com/rabbitmq/amqp091-go"
 )
 
@@ -74,10 +74,10 @@ func (c *connector) handle(message amqp.Delivery, ctx context.Context) {
 		result = router.NewResponse("", nil, fmt.Errorf("type(message property) must not be empty")).Marshal()
 	} else if message.MessageId == "" {
 		result = router.NewResponse("", nil, fmt.Errorf("message_id(message property) must not be empty")).Marshal()
-	} else {	
+	} else {
 		result = c.router.Route(message.Type, message.MessageId, message.Body)
 	}
-	
+
 	if err := c.producer.Publish(result, ctx); err != nil {
 		c.logger.Log(logger.ERROR, fmt.Sprintf("failed to publish result: %s: %s", string(result), err))
 		// nack
